requirements.txt
pandas
numpy
scikit-learn
lightgbm
tushare
plotly
streamlit
python-dotenv
